SGE plugin - bug fix in hibernate mapping file
git-svn-id: https://svn.broadinstitute.org/genepattern/trunk@9882 48a7a3f7-ad46-0410-b969-29950918272e

+++ gp-sge-queue/src/main/java/org/genepattern/server/executor/sge/JobRecorder.java
@@ -39,7 +39,7 @@ public class JobRecorder {
      * @param gpJobNo
      * @param sgeJob
      */
-    public void updateSgeJobRecord(long gpJobNo, BatchJob sgeJob) {
+    public void updateSgeJobRecord(int gpJobNo, BatchJob sgeJob) {
         JobSge jobRecord = new JobSge();
         jobRecord.setGpJobNo(gpJobNo);
         populateJobRecordFromSgeJob( jobRecord, sgeJob );
